Removed references to HID class in Mass Storage Host demo app, restructured config descriptor processing routine to better suit the demo.
Restructured DynAlloc library code to be easier to read, and faster in operation (after defrag, the first hole is used as it is the only hole, rather than searching for exact sized holes first, then one larger, etc).

Fixed HostChapter9.c - total number of bytes to transfer counter was a uint8_t in the send routine rather than a uint16_t, which was truncating the value causing lost data.

git-svn-id: http://lufa-lib.googlecode.com/svn@153 d5102386-fcda-11dd-9fdb-3debd5008f28

+++ MyUSB/Drivers/USB/LowLevel/HostChapter9.c
@@ -18,7 +18,7 @@ uint8_t USB_Host_SendControlRequest(uint8_t* DataBuffer)
 	bool     SOFGenEnabled  = USB_HOST_SOFGeneration_IsEnabled();
 	uint8_t  ReturnStatus   = HOST_SENDCONTROL_Sucessful;
 	uint8_t* HeaderByte     = (uint8_t*)&USB_HostRequest;
-	uint8_t  DataLen        = USB_HostRequest.DataLength;
+	uint16_t DataLen        = USB_HostRequest.DataLength;
 	uint16_t TimeoutCounter;
 
 	USB_HOST_SOFGeneration_Enable();
